Backend na mazání ingrediencí s validací

+++ server/abl/ingredient/delete-abl.js
@@ -1,21 +1,54 @@
 const ingredientDao = require('../../dao/file_storage/ingredient-dao');
+const recipeDao = require('../../dao/file_storage/recipe-dao');
 
 async function DeleteAbl(req, res) {
     try {
+
         const id = req.params.id;
+        let allRecipe = await recipeDao.getAllRecipes();
+        let matchedIngredient;
+
+
+        allRecipe.forEach(recipe => {
+
+            if (matchedIngredient === undefined) {
+                matchedIngredient = recipe.ingredients.find(e => e.id === id);
+            } else {
+                return;
+            }
+
+            
+
+
+        })
+
         
-        let deletedIngredient = await ingredientDao.deleteIngredient(id);
-        if (deletedIngredient) {
-            res.send(deletedIngredient);
+        console.log(matchedIngredient);
+
+
+
+        if (matchedIngredient === undefined) {
+
+            let deletedIngredient = await ingredientDao.deleteIngredient(id);
+
+            if (deletedIngredient) {
+                res.send(deletedIngredient);
+            } else {
+                res.status(404).send({
+                    errors: [`an ingredient with the id ${id} was not found`]
+                });
+            }
         } else {
-            res.status(404).send({
-                errors: [`an ingredient with the id ${id} was not found`]
+            console.error("Ingredient is currently used in a recipe, deletion failed");
+            res.status(500).send({
+                errors: ["Ingredient is currently used in a recipe, deletion failed"]
             });
+            return;
         }
     } catch (e) {
         console.error(e);
         res.status(500).send({
-            errors: [ e.message ]
+            errors: [e.message]
         });
     }
 }
